requirements.txt
streamlit
openai
requests
python-dotenv
gitpython
paramiko
